backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
motor==3.6.0
pydantic==2.9.0
pydantic-settings==2.6.0
python-dotenv==1.0.0
argon2-cffi==23.1.0
python-jose[cryptography]==3.3.0
python-multipart==0.0.9
email-validator==2.1.0.post1